Update lundarLanderBeta.py
I changed the way movement works so it is more like our origional idea. It is still modular.

Brother Manley's version was good but it had flaws. It would keep moving in the same direction when the key was released.  Along with that it would only apply force when you pressed the key (if you held it down it would not accelerate).

This also allows fuel to only be spent while the key is being pressed (thrusters on).

I only commented out Brother Manley's work that way we can still reference and possibly fall back on it if needed.