Took out "_data"
When the lander hit the ground, nothing would happen, and the program would say "missing positional argument "_data." So I took _data out of wall_hit_handler and now the walls disappear when the lander hits them. I don't know why it's not removing the player_sprite, but at least things as being removed like we wanted them to.